Added iBeacon JSON builder class - pulls from http://www.experiencepush.com/beacon_portal/beacon_request.php?id='senior_week'

+++ SeniorWeek2014/Beacon.m
@@ -238,7 +238,7 @@
  * @return void
  */
 -(void) track:(NSInteger)rssi{
-    [[[self.encounters objectAtIndex:0] distArray] addObject:[NSString stringWithFormat:@"%d", rssi]];
+    [[[self.encounters objectAtIndex:0] distArray] addObject:[NSString stringWithFormat:@"%ld", (long)rssi]];
 }
